Removed external dependencies; Stable UI with build capabilities
