+++ frontend/src-tauri/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "app"
+name = "meeting-minutes-frontend-app"
 version = "0.1.0"
-description = "A Tauri App"
-authors = ["you"]
-license = ""
-repository = ""
+description = "A Tauri App for meeting minutes"
+authors = ["Sujith S"]
+license = "MIT"
+repository = "https://github.com/Zackriya-Solutions/meeting-minutes"
 edition = "2021"
-rust-version = "1.77.2"
+rust-version = "1.77"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -16,18 +16,85 @@ crate-type = ["staticlib", "cdylib", "rlib"]
 
 [build-dependencies]
 tauri-build = { version = "2.0.2", features = [] }
+reqwest = { version = "0.11", features = ["blocking", "multipart", "json"] }
+which = "6.0.1"
 
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
+
+# Cross-platform audio capture
+cpal = { git = "https://github.com/Kree0/cpal.git", branch = "master" }
+
+# Wav encoding
+hound = "3.5"
+
+# Cli ! shouldn't be required if using as lib
+clap = { version = "4.3", features = ["derive"] }
+
+# Dates
+chrono = { version = "0.4.31", features = ["serde"] }
+
+
+# Log
 log = "0.4"
+tracing = "0.1.40"
+env_logger = "0.10"
+which = "6.0.1"
+
+# Bytes
+bytemuck = "1.16.1"
+
+# Async
+tokio = { version = "1.32.0", features = ["full", "tracing"] }
+
+
+reqwest = { version = "0.11", features = ["blocking", "multipart", "json"] }
+
+
+# crossbeam
+crossbeam = "0.8.4"
+dashmap = "6.1.0"
+
+# Directories
+dirs = "5.0.1"
+
+lazy_static = { version = "1.4.0" }
+realfft = "3.4.0"
+regex = "1.11.0"
+ndarray = "0.16"
+bytes = { version = "1.9.0", features = ["serde"] }
+
+esaxx-rs = "0.1.10"
+samplerate = { version = "0.2.4" }
+libsamplerate-sys = "0.1.10"
+
+symphonia = { version = "0.5.4", features = ["aac", "isomp4", "opt-simd"] }
+rand = "0.8.5"
+rubato = "0.15.0"
+
+ffmpeg-sidecar = { git = "https://github.com/nathanbabcock/ffmpeg-sidecar", branch = "main" }
+
+[target.'cfg(target_os = "macos")'.dependencies]
+once_cell = "1.17.1"
+objc = "0.2.7"
+
+
 tauri = { version = "2.1.1", features = [ "protocol-asset", "macos-private-api"] }
+
 tauri-plugin-log = { version = "2.0.0-rc", features = ["colored"] }
-tokio = { version = "1.32.0", features = ["full"] }
-screenpipe-audio = { path = "../../screenpipe/screenpipe-audio" }
 anyhow = "1.0"
 time = { version = "0.3", features = ["formatting"] }
 reqwest = { version = "0.11", features = ["multipart", "json"] }
 env_logger = "0.10.1"
 tauri-plugin-fs = "2.2.0"
 tauri-plugin-dialog = "2.0.0"
+
+[dev-dependencies]
+tempfile = "3.3.0"
+infer = "0.15"
+criterion = { version = "0.5.1", features = ["async_tokio"] }
+memory-stats = "1.0"
+strsim = "0.10.0"
+futures = "0.3.31"
+tracing-subscriber = "0.3.16"
