+++ frontend/src-tauri/build.rs
@@ -1,3 +1,5 @@
 fn main() {
+    #[cfg(target_os = "macos")]
+    println!("cargo:rustc-link-lib=framework=AVFoundation");
     tauri_build::build()
 }
