--- frontend/src-tauri/src/audio/Cargo.toml
@@ -1,121 +1,0 @@
-[package]
-name = "screenpipe-audio"
-version = { workspace = true }
-authors = { workspace = true }
-description = { workspace = true }
-repository = { workspace = true }
-license = { workspace = true }
-edition = { workspace = true }
-
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[build-dependencies]
-reqwest = { workspace = true }
-which = "7.0.0"
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-
-# Cross-platform audio capture
-cpal = { git = "https://github.com/Kree0/cpal.git", branch = "master" }
-
-# Wav encoding
-hound = "3.5"
-
-# Cli ! shouldn't be required if using as lib
-clap = { version = "4.3", features = ["derive"] }
-
-# Dates
-chrono = { version = "0.4.31", features = ["serde"] }
-
-# Local Embeddings + STT
-# TODO: feature metal, cuda, etc. see https://github.com/huggingface/candle/blob/main/candle-core/Cargo.toml
-candle = { workspace = true }
-candle-nn = { workspace = true }
-candle-transformers = { workspace = true }
-vad-rs = "0.1.4"
-tokenizers = { workspace = true }
-anyhow = "1.0.86"
-byteorder = "1.5.0"
-hf-hub = "0.3.2"
-# https://github.com/pdeljanov/Symphonia/tree/master?tab=readme-ov-file#optimizations
-symphonia = { version = "0.5.4", features = ["aac", "isomp4", "opt-simd"] }
-rand = "0.8.5"
-rubato = "0.15.0"
-
-# Log
-log = { workspace = true }
-tracing = { workspace = true }
-env_logger = "0.10"
-
-# Bytes
-bytemuck = "1.16.1"
-
-# Async
-tokio = { workspace = true }
-
-# Detect speech/silence
-webrtc-vad = "0.4.0"
-
-# Deepgram
-reqwest = { workspace = true }
-
-screenpipe-core = { path = "../screenpipe-core" }
-
-# crossbeam
-crossbeam = { workspace = true }
-
-# Directories
-dirs = "5.0.1"
-
-lazy_static = { version = "1.4.0" }
-realfft = "3.4.0"
-regex = "1.11.0"
-ndarray = "0.16"
-ort = "=2.0.0-rc.6"
-knf-rs = { git = "https://github.com/Neptune650/knf-rs.git" }
-ort-sys = "=2.0.0-rc.8"
-
-[target.'cfg(target_os = "windows")'.dependencies]
-ort = { version = "=2.0.0-rc.6", features = [
-  "download-binaries",
-  "copy-dylibs",
-  "directml",
-  "cuda",
-] }
-esaxx-rs = "0.1.10"
-samplerate = { version = "0.2.4" }
-libsamplerate-sys = "0.1.10"
-
-[target.'cfg(target_os = "macos")'.dependencies]
-once_cell = "1.17.1"
-objc = "0.2.7"
-
-[dev-dependencies]
-tempfile = "3.3.0"
-infer = "0.15"
-criterion = { workspace = true }
-memory-stats = "1.0"
-strsim = "0.10.0"
-futures = "0.3.31"
-tracing-subscriber = "0.3.16"
-[features]
-metal = ["candle/metal", "candle-nn/metal", "candle-transformers/metal"]
-cuda = ["candle/cuda", "candle-nn/cuda", "candle-transformers/cuda"]
-mkl = ["candle/mkl", "candle-nn/mkl", "candle-transformers/mkl"]
-
-[[bin]]
-name = "screenpipe-audio"
-path = "src/bin/screenpipe-audio.rs"
-
-[[bench]]
-name = "pcm_decode_benchmark"
-harness = false
-
-[[bench]]
-name = "record_and_transcribe_benchmark"
-harness = false
-
-[package.metadata.cargo-machete]
-ignored = ["ort-sys"]
