--- frontend/src-tauri/src/audio/README.md
@@ -1,4 +1,0 @@
-
-
-
-![image](https://github.com/user-attachments/assets/d27c6f28-7f68-44ba-9e8c-124019daf83a)
