+++ frontend/src-tauri/src/audio/core.rs
@@ -1,10 +1,9 @@
-use crate::audio_processing::audio_to_mono;
-use crate::AudioInput;
+use super::audio_processing::audio_to_mono; 
 use anyhow::{anyhow, Result};
 use cpal::traits::{DeviceTrait, HostTrait, StreamTrait};
 use cpal::StreamError;
 use lazy_static::lazy_static;
-use log::{debug, error, info, warn};
+use log::{ error, info, warn};
 use serde::{Deserialize, Serialize};
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 use std::sync::mpsc;
@@ -48,7 +47,7 @@ impl Default for AudioTranscriptionEngine {
     }
 }
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct DeviceControl {
     pub is_running: bool,
     pub is_paused: bool,
@@ -166,115 +165,6 @@ pub async fn get_device_and_config(
     Ok((cpal_audio_device, config))
 }
 
-pub async fn record_and_transcribe(
-    audio_stream: Arc<AudioStream>,
-    duration: Duration,
-    whisper_sender: crossbeam::channel::Sender<AudioInput>,
-    is_running: Arc<AtomicBool>,
-) -> Result<()> {
-    while is_running.load(Ordering::Relaxed) {
-        match run_record_and_transcribe(
-            audio_stream.clone(),
-            duration,
-            whisper_sender.clone(),
-            is_running.clone(),
-        )
-        .await
-        {
-            Ok(_) => {
-                // Normal shutdown
-                break;
-            }
-            Err(e) => {
-                if !is_running.load(Ordering::Relaxed) {
-                    // Normal shutdown
-                    break;
-                }
-
-                error!("record_and_transcribe error, restarting: {}", e);
-                // Add a small delay before restarting to prevent rapid restart loops
-                tokio::time::sleep(Duration::from_secs(1)).await;
-            }
-        }
-    }
-    Ok(())
-}
-
-async fn run_record_and_transcribe(
-    audio_stream: Arc<AudioStream>,
-    duration: Duration,
-    whisper_sender: crossbeam::channel::Sender<AudioInput>,
-    is_running: Arc<AtomicBool>,
-) -> Result<()> {
-    let mut receiver = audio_stream.subscribe().await;
-
-    info!(
-        "starting continuous recording for {} ({}s segments)",
-        audio_stream.device.to_string(),
-        duration.as_secs()
-    );
-
-    const OVERLAP_SECONDS: usize = 2;
-    let mut collected_audio = Vec::new();
-    let sample_rate = audio_stream.device_config.sample_rate().0 as usize;
-    let overlap_samples = OVERLAP_SECONDS * sample_rate;
-
-    while is_running.load(Ordering::Relaxed)
-        && !audio_stream.is_disconnected.load(Ordering::Relaxed)
-    {
-        let start_time = tokio::time::Instant::now();
-
-        while start_time.elapsed() < duration && is_running.load(Ordering::Relaxed) {
-            match tokio::time::timeout(Duration::from_millis(100), receiver.recv()).await {
-                Ok(Ok(chunk)) => {
-                    collected_audio.extend(chunk);
-                    LAST_AUDIO_CAPTURE.store(
-                        std::time::SystemTime::now()
-                            .duration_since(std::time::UNIX_EPOCH)
-                            .unwrap()
-                            .as_secs(),
-                        Ordering::Relaxed,
-                    );
-                }
-                Ok(Err(e)) => {
-                    error!("error receiving audio data: {}", e);
-                    return Err(anyhow!("Audio stream error: {}", e));
-                }
-                Err(_) => {} // Timeout, continue loop
-            }
-        }
-
-        if !collected_audio.is_empty() {
-            debug!("sending audio segment to audio model");
-            match whisper_sender.try_send(AudioInput {
-                data: Arc::new(collected_audio.clone()),
-                device: audio_stream.device.clone(),
-                sample_rate: audio_stream.device_config.sample_rate().0,
-                channels: audio_stream.device_config.channels(),
-            }) {
-                Ok(_) => {
-                    debug!("sent audio segment to audio model");
-                    if collected_audio.len() > overlap_samples {
-                        collected_audio =
-                            collected_audio.split_off(collected_audio.len() - overlap_samples);
-                    }
-                }
-                Err(e) => {
-                    if e.is_disconnected() {
-                        error!("whisper channel disconnected, restarting recording process");
-                        return Err(anyhow!("Whisper channel disconnected"));
-                    } else if e.is_full() {
-                        warn!("whisper channel full, dropping audio segment");
-                        tokio::time::sleep(Duration::from_millis(100)).await;
-                    }
-                }
-            }
-        }
-    }
-
-    info!("stopped recording for {}", audio_stream.device.to_string());
-    Ok(())
-}
 
 pub async fn list_audio_devices() -> Result<Vec<AudioDevice>> {
     let host = cpal::default_host();
