--- frontend/src-tauri/src/audio/src/pyannote/mod.rs
@@ -1,5 +1,0 @@
-pub mod embedding;
-pub mod identify;
-pub mod models;
-pub mod segment;
-pub mod session;
