--- frontend/src-tauri/src/audio/src/whisper.rs
@@ -1,422 +1,0 @@
-use anyhow::{Error as E, Result};
-use candle::{Device, IndexOp, Tensor};
-use candle_nn::{ops::softmax, VarBuilder};
-use candle_transformers::models::whisper::{self as m, Config};
-use hf_hub::{api::sync::Api, Repo, RepoType};
-use log::{debug, error, info};
-use rand::{distributions::Distribution, SeedableRng};
-use tokenizers::Tokenizer;
-
-#[derive(Clone)]
-pub struct WhisperModel {
-    pub model: Model,
-    pub tokenizer: Tokenizer,
-    pub device: Device,
-}
-
-impl WhisperModel {
-    pub fn new(engine: &crate::AudioTranscriptionEngine) -> Result<Self> {
-        debug!("Initializing WhisperModel");
-        let device = Device::new_metal(0).unwrap_or(Device::new_cuda(0).unwrap_or(Device::Cpu));
-        info!("device = {:?}", device);
-
-        debug!("Fetching model files");
-        let (config_filename, tokenizer_filename, weights_filename) = {
-            let api = Api::new()?;
-            let repo = match engine {
-                crate::AudioTranscriptionEngine::WhisperTiny => Repo::with_revision(
-                    "openai/whisper-tiny".to_string(),
-                    RepoType::Model,
-                    "main".to_string(),
-                ),
-                crate::AudioTranscriptionEngine::WhisperDistilLargeV3 => Repo::with_revision(
-                    "distil-whisper/distil-large-v3".to_string(),
-                    RepoType::Model,
-                    "main".to_string(),
-                ),
-                crate::AudioTranscriptionEngine::WhisperLargeV3Turbo => Repo::with_revision(
-                    "openai/whisper-large-v3-turbo".to_string(),
-                    RepoType::Model,
-                    "main".to_string(),
-                ),
-                _ => Repo::with_revision(
-                    "openai/whisper-large-v3-turbo".to_string(),
-                    RepoType::Model,
-                    "main".to_string(),
-                ),
-            };
-            let api_repo = api.repo(repo);
-            let config = api_repo.get("config.json")?;
-            let tokenizer = api_repo.get("tokenizer.json")?;
-            let model = api_repo.get("model.safetensors")?;
-            (config, tokenizer, model)
-        };
-
-        debug!("Parsing config and tokenizer");
-        let config: Config = serde_json::from_str(&std::fs::read_to_string(config_filename)?)?;
-        let tokenizer = Tokenizer::from_file(tokenizer_filename).map_err(E::msg)?;
-        // tokenizer.with_pre_tokenizer(PreT)
-        debug!("Loading model weights");
-        let vb =
-            unsafe { VarBuilder::from_mmaped_safetensors(&[weights_filename], m::DTYPE, &device)? };
-        let whisper = m::model::Whisper::load(&vb, config.clone())?;
-
-        let model = Model::Normal(whisper);
-
-        debug!("WhisperModel initialization complete");
-        Ok(Self {
-            model,
-            tokenizer,
-            device,
-        })
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum Model {
-    Normal(m::model::Whisper),
-    Quantized(m::quantized_model::Whisper),
-}
-
-impl Model {
-    pub fn config(&self) -> &Config {
-        match self {
-            Self::Normal(m) => &m.config,
-            Self::Quantized(m) => &m.config,
-        }
-    }
-
-    pub fn encoder_forward(&mut self, x: &Tensor, flush: bool) -> candle::Result<Tensor> {
-        match self {
-            Self::Normal(m) => m.encoder.forward(x, flush),
-            Self::Quantized(m) => m.encoder.forward(x, flush),
-        }
-    }
-
-    pub fn decoder_forward(
-        &mut self,
-        x: &Tensor,
-        xa: &Tensor,
-        flush: bool,
-    ) -> candle::Result<Tensor> {
-        match self {
-            Self::Normal(m) => m.decoder.forward(x, xa, flush),
-            Self::Quantized(m) => m.decoder.forward(x, xa, flush),
-        }
-    }
-
-    pub fn decoder_final_linear(&self, x: &Tensor) -> candle::Result<Tensor> {
-        match self {
-            Self::Normal(m) => m.decoder.final_linear(x),
-            Self::Quantized(m) => m.decoder.final_linear(x),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct DecodingResult {
-    tokens: Vec<u32>,
-    pub text: String,
-    avg_logprob: f64,
-    no_speech_prob: f64,
-    #[allow(dead_code)]
-    temperature: f64,
-    compression_ratio: f64,
-}
-
-#[derive(Debug, Clone)]
-pub struct Segment {
-    pub start: f64,
-    pub duration: f64,
-    pub dr: DecodingResult,
-}
-
-pub struct Decoder<'a> {
-    model: &'a mut Model,
-    rng: rand::rngs::StdRng,
-    timestamps: bool,
-    verbose: bool,
-    tokenizer: &'a Tokenizer,
-    suppress_tokens: Tensor,
-    sot_token: u32,
-    transcribe_token: u32,
-    eot_token: u32,
-    no_speech_token: u32,
-    no_timestamps_token: u32,
-    language_token: Option<u32>,
-}
-
-impl<'a> Decoder<'a> {
-    #[allow(clippy::too_many_arguments)]
-    pub fn new(
-        model: &'a mut Model,
-        tokenizer: &'a Tokenizer,
-        seed: u64,
-        device: &Device,
-        language_token: Option<u32>,
-        timestamps: bool,
-        verbose: bool,
-    ) -> Result<Self> {
-        let no_timestamps_token = token_id(&tokenizer, m::NO_TIMESTAMPS_TOKEN)?;
-        let suppress_tokens: Vec<f32> = (0..model.config().vocab_size as u32)
-            .map(|i| {
-                if model.config().suppress_tokens.contains(&i)
-                    || timestamps && i == no_timestamps_token
-                {
-                    f32::NEG_INFINITY
-                } else {
-                    0f32
-                }
-            })
-            .collect();
-        let suppress_tokens = Tensor::new(suppress_tokens.as_slice(), device)?;
-        let sot_token = token_id(&tokenizer, m::SOT_TOKEN)?;
-        let transcribe_token = token_id(&tokenizer, m::TRANSCRIBE_TOKEN)?;
-        let eot_token = token_id(&tokenizer, m::EOT_TOKEN)?;
-        let no_speech_token = m::NO_SPEECH_TOKENS
-            .iter()
-            .find_map(|token| token_id(&tokenizer, token).ok());
-        let no_speech_token = match no_speech_token {
-            None => anyhow::bail!("unable to find any non-speech token"),
-            Some(n) => n,
-        };
-
-        Ok(Self {
-            model,
-            rng: rand::rngs::StdRng::seed_from_u64(seed),
-            tokenizer,
-            timestamps,
-            verbose,
-            suppress_tokens,
-            sot_token,
-            transcribe_token,
-            eot_token,
-            no_speech_token,
-            language_token,
-            no_timestamps_token,
-        })
-    }
-
-    fn decode(&mut self, mel: &Tensor, t: f64) -> Result<DecodingResult> {
-        let audio_features = self.model.encoder_forward(mel, true)?;
-        if self.verbose {
-            info!("audio features: {:?}", audio_features.dims());
-        }
-        let sample_len = self.model.config().max_target_positions / 2;
-        let mut no_speech_prob = f64::NAN;
-        let mut tokens = vec![self.sot_token];
-        if let Some(language_token) = self.language_token {
-            tokens.push(language_token);
-        }
-        tokens.push(self.transcribe_token);
-
-        if !self.timestamps {
-            tokens.push(self.no_timestamps_token);
-        }
-
-        let mut sum_logprob = 0f64;
-        let mut last_token_was_timestamp = false;
-
-        let mut token_history = Vec::new(); // Track recent tokens
-
-        for i in 0..sample_len {
-            let tokens_t = Tensor::new(tokens.as_slice(), mel.device())?;
-            let tokens_t = tokens_t.unsqueeze(0)?;
-            let ys = self
-                .model
-                .decoder_forward(&tokens_t, &audio_features, i == 0)?;
-
-            if i == 0 {
-                let logits = self.model.decoder_final_linear(&ys.i(..1)?)?.i(0)?.i(0)?;
-                no_speech_prob = softmax(&logits, 0)?
-                    .i(self.no_speech_token as usize)?
-                    .to_scalar::<f32>()? as f64;
-            }
-
-            let (_, seq_len, _) = ys.dims3()?;
-            let logits = self
-                .model
-                .decoder_final_linear(&ys.i((..1, seq_len - 1..))?)?
-                .i(0)?
-                .i(0)?;
-
-            let logits = logits.broadcast_add(&self.suppress_tokens)?;
-
-            let logits = if last_token_was_timestamp {
-                let mask = Tensor::zeros_like(&logits)?;
-                let eot_mask = mask.get(self.eot_token as usize)?;
-                logits.broadcast_add(&eot_mask)?
-            } else {
-                logits
-            };
-
-            // Apply repetition penalty
-            let mut logits_v: Vec<f32> = logits.to_vec1()?;
-            apply_repetition_penalty(&mut logits_v, &token_history, 1.0); // Adjust penalty as needed
-
-            let next_token = if t > 0f64 {
-                let logits_tensor = Tensor::new(logits_v.as_slice(), logits.device())?;
-                let scaled_logits = (&logits_tensor / t)?;
-                let prs = softmax(&scaled_logits, 0)?;
-                let prs_vec: Vec<f32> = prs.to_vec1()?;
-                let distr = rand::distributions::WeightedIndex::new(&prs_vec)?;
-                distr.sample(&mut self.rng) as u32
-            } else {
-                logits_v
-                    .iter()
-                    .enumerate()
-                    .max_by(|(_, u), (_, v)| u.total_cmp(v))
-                    .map(|(i, _)| i as u32)
-                    .unwrap()
-            };
-
-            tokens.push(next_token);
-            token_history.push(next_token); // Add to history
-
-            let prob = softmax(&logits, candle::D::Minus1)?
-                .i(next_token as usize)?
-                .to_scalar::<f32>()? as f64;
-
-            sum_logprob += prob.ln();
-
-            if next_token == self.eot_token
-                || tokens.len() > self.model.config().max_target_positions
-            {
-                break;
-            }
-
-            last_token_was_timestamp = next_token > self.no_timestamps_token;
-        }
-
-        let text = self.tokenizer.decode(&tokens, true).map_err(E::msg)?;
-        let avg_logprob = sum_logprob / tokens.len() as f64;
-
-        Ok(DecodingResult {
-            tokens,
-            text,
-            avg_logprob,
-            no_speech_prob,
-            temperature: t,
-            compression_ratio: f64::NAN,
-        })
-    }
-
-    fn decode_with_fallback(&mut self, segment: &Tensor) -> Result<DecodingResult> {
-        for (i, &t) in m::TEMPERATURES.iter().enumerate() {
-            let dr: Result<DecodingResult> = self.decode(segment, t);
-            if i == m::TEMPERATURES.len() - 1 {
-                return dr;
-            }
-            match dr {
-                Ok(dr) => {
-                    let needs_fallback = dr.compression_ratio > m::COMPRESSION_RATIO_THRESHOLD
-                        || dr.avg_logprob < m::LOGPROB_THRESHOLD;
-                    if !needs_fallback || dr.no_speech_prob > m::NO_SPEECH_THRESHOLD {
-                        return Ok(dr);
-                    }
-                }
-                Err(err) => {
-                    error!("Error running at {t}: {err}")
-                }
-            }
-        }
-        unreachable!()
-    }
-
-    pub fn run(&mut self, mel: &Tensor) -> Result<Vec<Segment>> {
-        let (_, _, content_frames) = mel.dims3()?;
-        let mut seek = 0;
-        let mut segments = vec![];
-        while seek < content_frames {
-            let start = std::time::Instant::now();
-            let time_offset = (seek * m::HOP_LENGTH) as f64 / m::SAMPLE_RATE as f64;
-            let segment_size = usize::min(content_frames - seek, m::N_FRAMES);
-            let mel_segment = mel.narrow(2, seek, segment_size)?;
-            let segment_duration = (segment_size * m::HOP_LENGTH) as f64 / m::SAMPLE_RATE as f64;
-            let dr = self.decode_with_fallback(&mel_segment)?;
-            seek += segment_size;
-            if dr.no_speech_prob > m::NO_SPEECH_THRESHOLD && dr.avg_logprob < m::LOGPROB_THRESHOLD {
-                info!("no speech detected, skipping {seek} {dr:?}");
-                continue;
-            }
-            let segment = Segment {
-                start: time_offset,
-                duration: segment_duration,
-                dr,
-            };
-            if self.timestamps {
-                // info!(
-                //     "{:.1}s -- {:.1}s",
-                //     segment.start,
-                //     segment.start + segment.duration,
-                // );
-                let mut tokens_to_decode = vec![];
-                let mut prev_timestamp_s = 0f32;
-                for &token in segment.dr.tokens.iter() {
-                    if token == self.sot_token || token == self.eot_token {
-                        continue;
-                    }
-                    if token > self.no_timestamps_token {
-                        let timestamp_s = (token - self.no_timestamps_token + 1) as f32 / 50.;
-                        if !tokens_to_decode.is_empty() {
-                            let text = self
-                                .tokenizer
-                                .decode(&tokens_to_decode, true)
-                                .map_err(E::msg)?;
-                            info!("  {:.1}s-{:.1}s: {}", prev_timestamp_s, timestamp_s, text);
-                            tokens_to_decode.clear()
-                        }
-                        prev_timestamp_s = timestamp_s;
-                    } else {
-                        tokens_to_decode.push(token)
-                    }
-                }
-                if !tokens_to_decode.is_empty() {
-                    let text = self
-                        .tokenizer
-                        .decode(&tokens_to_decode, true)
-                        .map_err(E::msg)?;
-                    if !text.is_empty() {
-                        // info!("  {:.1}s-...: {}", prev_timestamp_s, text);
-                    }
-                    tokens_to_decode.clear()
-                }
-            } else {
-                // info!(
-                //     "{:.1}s -- {:.1}s: {}",
-                //     segment.start,
-                //     segment.start + segment.duration,
-                //     segment.dr.text,
-                // )
-            }
-            if self.verbose {
-                info!("{seek}: {segment:?}, in {:?}", start.elapsed());
-            }
-            segments.push(segment)
-        }
-        Ok(segments)
-    }
-}
-pub fn token_id(tokenizer: &Tokenizer, token: &str) -> candle::Result<u32> {
-    match tokenizer.token_to_id(token) {
-        None => candle::bail!("no token-id for {token}"),
-        Some(id) => Ok(id),
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum Task {
-    Transcribe,
-    #[allow(dead_code)]
-    Translate,
-}
-
-fn apply_repetition_penalty(logits: &mut [f32], token_history: &[u32], penalty: f32) {
-    for &token in token_history {
-        if let Some(logit) = logits.get_mut(token as usize) {
-            *logit -= penalty;
-        }
-    }
-}
-
-// ... (keep any other helper functions or structs that are specific to Whisper)
